+++ utils/utils.go
@@ -6,11 +6,11 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-func FmtNodeId(nodeid uint32) string {
+func FmtNodeId(nodeid int64) string {
 	return fmt.Sprintf("0x%06X", nodeid)
 }
 
-func FmtNodeIdHass(nodeid uint32) string {
+func FmtNodeIdHass(nodeid int64) string {
 	return fmt.Sprintf("127.%d.%d.%d", (nodeid>>16)&0xFF, (nodeid>>8)&0xFF, nodeid&0xFF)
 }
